Move PROTO_VERSION from veles.server to veles.proto
Importing veles.server from veles.scli breaks Python 2 for obvious reasons.

Reported by mkow.
